OneDrive/Desktop/RevivalUK/Q1venv/source/requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
boto3==1.16.25
botocore==1.19.25
Brotli==1.0.9
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2
daphne==3.0.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-extensions==3.0.9
django-filter==2.4.0
django-heroku==0.3.1
django-neomodel==0.0.4
django-request==1.5.6
django-storages==1.10.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
docker==4.3.1
drf-yasg==1.17.1
ecdsa==0.14.1
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
graphviz==0.14.2
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
msgpack==1.0.0
neo4j-driver==4.1.1
neobolt==1.7.17
neomodel==4.0.1
oauthlib==3.1.0
packaging==20.4
plotly==4.13.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
python-jose==3.2.0
python3-openid==3.2.0
pytz==2020.1
pywin32==227
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
service-identity==18.1.0
Shapely==1.7.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.11
websocket-client==0.57.0
Werkzeug==1.0.1
whitenoise==5.2.0
zope.interface==5.2.0
